requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
Django==5.0.2
django-cors-headers==4.3.1
django-countries==7.5.1
django-extensions==3.2.3
djangorestframework==3.14.0
exceptiongroup==1.2.0
googlesearch-python==1.2.3
h11==0.14.0
idna==3.6
lxml==5.1.0
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
selenium==4.17.2
Selenium-Screenshot==2.1.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
webdriver-manager==4.0.1
wsproto==1.2.0
